+++ prove/Develop02/journal.csv
@@ -10,3 +10,6 @@ i like pineapple
 Monday, October 23, 2023
 
 I like to ride my bicyle
+Thursday, January 18, 2024
+What was the best part of my day?
+programming
